requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.3.0
pytesseract==0.2.5
requests==2.20.1
tesseract==0.1.3
urllib3==1.24.1
Werkzeug==0.14.1
